refactor(cache): remove local cache in redis
