+++ internal/services/cache/init.go
@@ -8,9 +8,7 @@ import (
 	"os"
 	"strings"
 	"sync"
-	"time"
 
-	"github.com/go-redis/redis/v8"
 	"github.com/rs/zerolog/log"
 )
 
@@ -44,36 +42,6 @@ var (
 	mu          sync.RWMutex
 )
 
-// getRedisOptions returns Redis configuration optimized for the current environment
-func getRedisOptions(addr string) *redis.Options {
-	isDev := os.Getenv("GIN_MODE") != "release"
-
-	// Base configuration optimized for single user
-	opts := &redis.Options{
-		Addr:            addr,
-		MinIdleConns:    1,
-		MaxRetries:      RetryAttempts,
-		MinRetryBackoff: RetryDelay,
-		MaxRetryBackoff: time.Second,
-		// Reduced pool size for single user scenario
-		PoolSize:     3,
-		MaxConnAge:   5 * time.Minute,
-		IdleTimeout:  30 * time.Second,
-		ReadTimeout:  DefaultTimeout,
-		WriteTimeout: DefaultTimeout,
-		PoolTimeout:  DefaultTimeout,
-	}
-
-	if isDev {
-		// Even smaller settings for development
-		opts.PoolSize = 2
-		opts.MaxConnAge = 30 * time.Second
-		opts.IdleTimeout = 15 * time.Second
-	}
-
-	return opts
-}
-
 // getCacheType determines which cache implementation to use based on environment
 func getCacheType() CacheType {
 	cacheType := os.Getenv("CACHE_TYPE")
@@ -99,55 +67,13 @@ func getCacheType() CacheType {
 // createCache creates a new cache instance
 func createCache(ctx context.Context, cfg Config) (Store, error) {
 	cacheType := getCacheType()
-	var err error
 
 	switch cacheType {
 	case CacheTypeRedis:
 		if cfg.RedisAddr == "" {
 			return NewMemoryStore(ctx, cfg.DataDir), nil
 		}
-
-		isDev := os.Getenv("GIN_MODE") != "release"
-		opts := getRedisOptions(cfg.RedisAddr)
-
-		timeout := DefaultTimeout
-		if isDev {
-			timeout = 2 * time.Second
-		}
-
-		timeoutCtx, cancel := context.WithTimeout(ctx, timeout)
-		defer cancel()
-
-		client := redis.NewClient(opts)
-
-		err = client.Ping(timeoutCtx).Err()
-		if err != nil {
-			if client != nil {
-				client.Close()
-			}
-			if os.Getenv("CACHE_TYPE") == "redis" {
-				log.Error().Err(err).Str("addr", opts.Addr).Msg("Failed to connect to explicitly configured Redis, falling back to memory cache")
-			}
-			return NewMemoryStore(ctx, cfg.DataDir), err
-		}
-
-		storeCtx, storeCancel := context.WithCancel(ctx)
-		store := &RedisStore{
-			client: client,
-			local: &LocalCache{
-				items: make(map[string]*localCacheItem),
-			},
-			ctx:    storeCtx,
-			cancel: storeCancel,
-		}
-
-		store.wg.Add(1)
-		go func() {
-			defer store.wg.Done()
-			store.localCacheCleanup()
-		}()
-
-		return store, nil
+		return NewRedisStore(cfg), nil
 
 	case CacheTypeMemory:
 		return NewMemoryStore(ctx, cfg.DataDir), nil
